+++ voxygen/src/hud/minimap.rs
@@ -383,6 +383,7 @@ widget_ids! {
         mmap_frame_2,
         mmap_frame_bg,
         mmap_location,
+        mmap_coordinates,
         mmap_button,
         mmap_plus,
         mmap_minus,
@@ -476,8 +477,16 @@ impl Widget for MiniMap<'_> {
         let mut events = Vec::new();
 
         let widget::UpdateArgs { state, ui, .. } = args;
+        let colored_player_marker = self
+            .global_state
+            .settings
+            .interface
+            .minimap_colored_player_marker;
         let mut zoom = self.global_state.settings.interface.minimap_zoom;
-        const SCALE: f64 = 1.5; // TODO Make this a setting
+        let mut scale = self.global_state.settings.interface.minimap_scale;
+        if scale <= 0.0 {
+            scale = 1.5;
+        }
         let show_minimap = self.global_state.settings.interface.minimap_show;
         let is_facing_north = self.global_state.settings.interface.minimap_face_north;
         let show_topo_map = self.global_state.settings.interface.map_show_topo_map;
@@ -487,20 +496,21 @@ impl Widget for MiniMap<'_> {
         } else {
             self.ori
         };
+        let map_size = Vec2::new(170.0 * scale, 170.0 * scale);
 
         if show_minimap {
             Image::new(self.imgs.mmap_frame)
-                .w_h(174.0 * SCALE, 190.0 * SCALE)
+                .w_h(174.0 * scale, 190.0 * scale)
                 .top_right_with_margins_on(ui.window, 5.0, 5.0)
                 .color(Some(UI_MAIN))
                 .set(state.ids.mmap_frame, ui);
             Image::new(self.imgs.mmap_frame_2)
-                .w_h(174.0 * SCALE, 190.0 * SCALE)
+                .w_h(174.0 * scale, 190.0 * scale)
                 .middle_of(state.ids.mmap_frame)
                 .color(Some(UI_HIGHLIGHT_0))
                 .set(state.ids.mmap_frame_2, ui);
-            Rectangle::fill_with([170.0 * SCALE, 170.0 * SCALE], color::TRANSPARENT)
-                .mid_top_with_margin_on(state.ids.mmap_frame_2, 18.0 * SCALE)
+            Rectangle::fill_with([170.0 * scale, 170.0 * scale], color::TRANSPARENT)
+                .mid_top_with_margin_on(state.ids.mmap_frame_2, 18.0 * scale)
                 .set(state.ids.mmap_frame_bg, ui);
 
             // Map size in chunk coords
@@ -541,7 +551,7 @@ impl Widget for MiniMap<'_> {
             let can_zoom_out = zoom > min_zoom;
 
             if Button::image(self.imgs.mmap_minus)
-                .w_h(16.0 * SCALE, 18.0 * SCALE)
+                .w_h(16.0 * scale, 18.0 * scale)
                 .hover_image(self.imgs.mmap_minus_hover)
                 .press_image(self.imgs.mmap_minus_press)
                 .top_left_with_margins_on(state.ids.mmap_frame, 0.0, 0.0)
@@ -558,7 +568,7 @@ impl Widget for MiniMap<'_> {
             }
 
             if Button::image(self.imgs.mmap_plus)
-                .w_h(18.0 * SCALE, 18.0 * SCALE)
+                .w_h(18.0 * scale, 18.0 * scale)
                 .hover_image(self.imgs.mmap_plus_hover)
                 .press_image(self.imgs.mmap_plus_press)
                 .right_from(state.ids.mmap_minus, 0.0)
@@ -579,7 +589,7 @@ impl Widget for MiniMap<'_> {
             } else {
                 self.imgs.mmap_north
             })
-            .w_h(18.0 * SCALE, 18.0 * SCALE)
+            .w_h(18.0 * scale, 18.0 * scale)
             .hover_image(if is_facing_north {
                 self.imgs.mmap_north_press_hover
             } else {
@@ -620,8 +630,6 @@ impl Widget for MiniMap<'_> {
                 ],
                 [w_src, h_src],
             );
-
-            let map_size = Vec2::new(170.0 * SCALE, 170.0 * SCALE);
 
             // Map Image
             // Map Layer Images
@@ -915,9 +923,15 @@ impl Widget for MiniMap<'_> {
             // Indicator
             let ind_scale = 0.4;
             let ind_rotation = if is_facing_north {
-                self.rot_imgs.indicator_mmap_small.target_north
+                if colored_player_marker {
+                    self.rot_imgs.indicator_mmap_colored.target_north
+                } else {
+                    self.rot_imgs.indicator_mmap.target_north
+                }
+            } else if colored_player_marker {
+                self.rot_imgs.indicator_mmap_colored.none
             } else {
-                self.rot_imgs.indicator_mmap_small.none
+                self.rot_imgs.indicator_mmap.none
             };
             Image::new(ind_rotation)
                 .middle_of(state.ids.map_layers[0])
@@ -956,7 +970,7 @@ impl Widget for MiniMap<'_> {
             }
         } else {
             Image::new(self.imgs.mmap_frame_closed)
-                .w_h(174.0 * SCALE, 18.0 * SCALE)
+                .w_h(174.0 * scale, 18.0 * scale)
                 .color(Some(UI_MAIN))
                 .top_right_with_margins_on(ui.window, 0.0, 5.0)
                 .set(state.ids.mmap_frame, ui);
@@ -967,7 +981,7 @@ impl Widget for MiniMap<'_> {
         } else {
             self.imgs.mmap_closed
         })
-        .w_h(18.0 * SCALE, 18.0 * SCALE)
+        .w_h(18.0 * scale, 18.0 * scale)
         .hover_image(if show_minimap {
             self.imgs.mmap_open_hover
         } else {
@@ -995,15 +1009,20 @@ impl Widget for MiniMap<'_> {
                 // Count characters in the name to avoid clipping with the name display
                 if let Some(name) = chunk.meta().name() {
                     let name_len = name.chars().count();
+                    let pos = map_size / 2.0 + 3.0;
                     Text::new(name)
-                        .mid_top_with_margin_on(state.ids.mmap_frame, match name_len {
-                            15..=30 => 4.0,
-                            _ => 2.0,
-                        })
+                        .align_middle_x_of(state.ids.mmap_frame)
+                        .y_position_relative_to(
+                            state.ids.mmap_frame,
+                            position::Relative::Scalar(pos.y),
+                        )
                         .font_size(self.fonts.cyri.scale(match name_len {
-                            0..=15 => 18,
-                            16..=30 => 14,
-                            _ => 14,
+                            0..=5 => 12 + (4.0 * scale).round() as u32,
+                            6..=10 => 10 + (4.0 * scale).round() as u32,
+                            11..=15 => 8 + (4.0 * scale).round() as u32,
+                            16..=20 => 6 + (4.0 * scale).round() as u32,
+                            21..=25 => 4 + (4.0 * scale).round() as u32,
+                            _ => 2 + (4.0 * scale).round() as u32,
                         }))
                         .font_id(self.fonts.cyri.conrod_id)
                         .color(TEXT_COLOR)