+++ CHANGELOG.md
@@ -50,6 +50,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New LoD models for all five AirshipDock plots.
 - Structures now support a `Choice()` block to allow for randomized spawns.
 - Added `DiscardTransparency` experimental shader.
+- Added Minimap options for Scale, Show Coordinates and Colored Player Marker.
 - Simple quest system
 - Quest: Escort a merchant
 - Quest: Slay a monster